Update the stored procedure helpers in YCSB
