+++ src/main/java/org/vanilladb/bench/server/procedure/ycsb/YcsbSchemaBuilderProc.java
@@ -1,19 +1,19 @@
 package org.vanilladb.bench.server.procedure.ycsb;
 
-import org.vanilladb.bench.server.param.ycsb.YcsbSchemaBuilderProcParamHelper;
+import org.vanilladb.bench.server.param.ycsb.YcsbSchemaBuilderSpParamHelper;
 import org.vanilladb.bench.server.procedure.StoredProcedureHelper;
 import org.vanilladb.core.sql.storedprocedure.StoredProcedure;
 import org.vanilladb.core.storage.tx.Transaction;
 
-public class YcsbSchemaBuilderProc extends StoredProcedure<YcsbSchemaBuilderProcParamHelper> {
+public class YcsbSchemaBuilderProc extends StoredProcedure<YcsbSchemaBuilderSpParamHelper> {
 
 	public YcsbSchemaBuilderProc() {
-		super(new YcsbSchemaBuilderProcParamHelper());
+		super(new YcsbSchemaBuilderSpParamHelper());
 	}
 
 	@Override
 	protected void executeSql() {
-		YcsbSchemaBuilderProcParamHelper paramHelper = getParamHelper();
+		YcsbSchemaBuilderSpParamHelper paramHelper = getParamHelper();
 		Transaction tx = getTransaction();
 		for (String sql : paramHelper.getTableSchemas())
 			StoredProcedureHelper.executeUpdate(sql, tx);
